Change weight system in map.
By default, when icon_weight is None, we put a default weight of 1
instead of 0. Because weights are used to draw map circles around
markers, we have to disable explicitely the drawing of circles if
icon_weight is None, we can no longer rely on the value.
This commit is useful when drawing the globe, because we can use
weights to compute heights of bars.